Update testing_bot
There are a lot of changes to the code. I think that you only need to change how the bot is going to get the list of users and maybe how it's going to send the message (the functions start() and message(), but the algorithm seems to be working as intended now. I don't think that you need to change anything. The only thing that you might need to change is the the_check variable. I didn't delete it, but I don't think it will affect the code if you remove it.